+++ analysis/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
+## [unreleased] (Added 🚀 | Changed | Removed 🗑 | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
+
+### Changed
+
+- Removed the top level from the output structure of the SonarImporter to match the other parsers and importers [3912](https://github.com/MaibornWolff/codecharta/pull/3912)
+
 ## [1.130.0] - 2024-12-17 (Added 🚀 | Changed | Removed 🗑 | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
 ### Added 🚀