+++ analysis/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased] (Added 🚀 | Changed | Removed 🗑 | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
+### Changed
+
+- Removed the top level from the output structure of the SonarImporter to match the other parsers and importers [3912](https://github.com/MaibornWolff/codecharta/pull/3912)
+
 ## [1.130.0] - 2024-12-17 (Added 🚀 | Changed | Removed 🗑 | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
 ### Added 🚀